Optimize parity and reduce cost of bit_length operations
- switch bit_length from 64-bit to 32-bit to avoid slow 64-bit division
- simplify parity logic (remove branch, use bitwise parity check)
- align 32-bit parity path with updated implementation
- tune reorder block_size to 32 for better cache locality

+++ include/sbd/chemistry/basic/mult.h
@@ -19,7 +19,7 @@ namespace sbd
 template <typename ElemT>
 class MultBase {
 protected:
-    size_t bit_length_;
+    uint32_t bit_length_;
     size_t norbs_;
     size_t D_size_;      // the vector length of a full (i.e., alpha + beta) determinant
     size_t D_half_size_; // the vector length of a half (i.e., alpha or beta) determinant
@@ -36,14 +36,14 @@ protected:
 public:
     MultBase() {}
 
-    MultBase(size_t bit_length, size_t norbs, MPI_Comm h_comm, MPI_Comm b_comm, MPI_Comm t_comm)
+    MultBase(uint32_t bit_length, size_t norbs, MPI_Comm h_comm, MPI_Comm b_comm, MPI_Comm t_comm)
         : bit_length_(bit_length), norbs_(norbs), h_comm_(h_comm), b_comm_(b_comm), t_comm_(t_comm)
     {
         D_size_ = (2 * norbs + bit_length - 1) / bit_length;
         D_half_size_ = (norbs + bit_length - 1) / bit_length;
     }
 
-    inline size_t bit_length(void) const
+    inline uint32_t bit_length(void) const
     {
         return bit_length_;
     }
